requirements.txt
Flask 
openai 
python-dotenv 
moviepy 
pydub 
gTTS 
requests 
beautifulsoup4 
pytrends 
Pillow 
